backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
Werkzeug==3.0.1
openai==1.12.0
Pillow==10.2.0
PyPDF2==3.0.1
python-multipart==0.0.9
bcrypt==4.1.2